+++ projects/ui/src/lib/components/tooltip/index.ts
@@ -1,2 +1,4 @@
 export { KuiTooltipDirective } from './kui-tooltip.directive';
 export type { KuiTooltipPlacement } from './kui-tooltip-placement.type';
+export type { KuiTooltipTrigger } from './kui-tooltip-trigger.type';
+export { KuiTooltipTriggerType } from './kui-tooltip-trigger.type';